+++ src/lib/utils-common/general.ts
@@ -1,5 +1,3 @@
-import { DefaultMap } from "./map.js";
-
 /**
  * This type provides a flag that can be used to turn off more lax overloads intended for
  * plugin use only to catch type errors in the TypeDoc codebase. The prepublishOnly npm
@@ -37,88 +35,12 @@ export type IfInternal<T, F> = InternalOnly extends true ? T : F;
 export type NeverIfInternal<T> = IfInternal<never, T>;
 
 /**
- * Resolves a string type into a union of characters, `"ab"` turns into `"a" | "b"`.
- */
-export type Chars<T extends string> = T extends `${infer C}${infer R}` ? C | Chars<R> :
-    never;
-
-/**
  * Utility to help type checking ensure that there is no uncovered case.
  */
 export function assertNever(x: never): never {
     throw new Error(
         `Expected handling to cover all possible cases, but it didn't cover: ${JSON.stringify(x)}`,
     );
-}
-
-// From MDN
-export function escapeRegExp(s: string) {
-    return s.replace(/[.*+?^${}()|[\]\\]/g, "\\$&"); // $& means the whole matched string
-}
-
-export function dedent(text: string) {
-    const lines = text.split(/\r?\n/);
-    while (lines.length && lines[0].search(/\S/) === -1) {
-        lines.shift();
-    }
-    while (lines.length && lines[lines.length - 1].search(/\S/) === -1) {
-        lines.pop();
-    }
-
-    const minIndent = lines.reduce(
-        (indent, line) => line.length ? Math.min(indent, line.search(/\S/)) : indent,
-        Infinity,
-    );
-
-    return lines.map((line) => line.substring(minIndent)).join("\n");
-}
-
-// Based on https://en.wikipedia.org/wiki/Levenshtein_distance#Iterative_with_two_matrix_rows
-// Slightly modified for improved match results for options
-export function editDistance(s: string, t: string): number {
-    if (s.length < t.length) return editDistance(t, s);
-
-    let v0 = Array.from({ length: t.length + 1 }, (_, i) => i);
-    let v1 = Array.from({ length: t.length + 1 }, () => 0);
-
-    for (let i = 0; i < s.length; i++) {
-        v1[0] = i + 1;
-
-        for (let j = 0; j < s.length; j++) {
-            const deletionCost = v0[j + 1] + 1;
-            const insertionCost = v1[j] + 1;
-            let substitutionCost: number;
-            if (s[i] === t[j]) {
-                substitutionCost = v0[j];
-            } else if (s[i]?.toUpperCase() === t[j]?.toUpperCase()) {
-                substitutionCost = v0[j] + 1;
-            } else {
-                substitutionCost = v0[j] + 3;
-            }
-
-            v1[j + 1] = Math.min(deletionCost, insertionCost, substitutionCost);
-        }
-
-        [v0, v1] = [v1, v0];
-    }
-
-    return v0[t.length];
-}
-
-export function getSimilarValues(values: Iterable<string>, compareTo: string) {
-    const results = new DefaultMap<number, string[]>(() => []);
-    let lowest = Infinity;
-    for (const name of values) {
-        const distance = editDistance(compareTo, name);
-        lowest = Math.min(lowest, distance);
-        results.get(distance).push(name);
-    }
-
-    // Experimenting a bit, it seems an edit distance of 3 is roughly the
-    // right metric for relevant "similar" results without showing obviously wrong suggestions
-    return results
-        .get(lowest)
-        .concat(results.get(lowest + 1), results.get(lowest + 2));
 }
 
 export function NonEnumerable(