+++ src/test/utils/options/readers/tsconfig.test.ts
@@ -1,8 +1,7 @@
 import { basename, join } from "path";
 import { deepStrictEqual as equal } from "assert";
 
-import { TSConfigReader } from "../../../../lib/utils/options/readers/index.js";
-import { Logger, Options } from "../../../../lib/utils/index.js";
+import { Logger, normalizePath, Options, TSConfigReader } from "#node-utils";
 import { TestLogger } from "../../../TestLogger.js";
 import { type Project, tempdirProject } from "@typestrong/fs-fixture-builder";
 import { tmpdir } from "os";
@@ -23,7 +22,7 @@ describe("Options - TSConfigReader", () => {
             options.reset();
         }
         logger.reset();
-        options.setValue("tsconfig", project.cwd);
+        options.setValue("tsconfig", normalizePath(project.cwd));
         project.addFile("temp.ts", "export {}");
         project.write();
         await options.read(logger);
@@ -38,7 +37,7 @@ describe("Options - TSConfigReader", () => {
 
         options.setValue(
             "tsconfig",
-            join(tmpdir(), "typedoc/does-not-exist.json"),
+            normalizePath(join(tmpdir(), "typedoc/does-not-exist.json")),
         );
         await options.read(logger);
         logger.expectMessage("error: *");
@@ -93,7 +92,7 @@ describe("Options - TSConfigReader", () => {
 
         options.setValue(
             "tsconfig",
-            join(fileURLToPath(import.meta.url), "../data/does_not_exist.json"),
+            normalizePath(join(fileURLToPath(import.meta.url), "../data/does_not_exist.json")),
         );
         options.addReader(new TSConfigReader());
         await options.read(logger);